+++ testdata/disjunct_struct_error.txtar
