+++ testdata/namelessNumericserror.txtar
