+++ testdata/orEnum.txtar
