+++ testdata/structEnumerror.txtar
